FirstLastNames' files updated and custom link enabled
FirstLastName plugin's 2012 registration view files' code has been
replaced with the updated 2014 view files' code from the 2.1 release of
Vanilla, except for the code for first and last name fields which was
changed to match the style of the new registration pages.
Additionally,  ReadMe and license files were added, and the link to the
Terms of Service page was changed in all registration view files to a
Terms of Use link that links to a user specified page.